fix member functions/types

+++ src/mon_sorter/mon_sorter.py
@@ -1,5 +1,5 @@
 
 class monomer:
     def __init__(self, distance: float, idx: int) -> None:
-        self.distance: int = distance
+        self.distance: float = distance
         self.idx: int = idx
